requirements.txt
matplotlib
tqdm
pandas
numpy
ipython
ipykernel
opencv-python
Pillow
scikit-image
pyyaml==5.1
tensorflow==2.7.0
mtcnn